requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose
cryptography
pydantic
mysql-connector-python
email-validator
